change author header

+++ src/main/java/com/dena/platform/core/feature/datastore/DenaDataStore.java
@@ -1,6 +1,6 @@
 package com.dena.platform.core.feature.datastore;
 
-import com.dena.platform.restapi.EntityDTO;
+import com.dena.platform.core.EntityDTO;
 
 /**
  * @author Javad Alimohammadi [<bs.alimohammadi@yahoo.com>]
@@ -9,6 +9,6 @@ import com.dena.platform.restapi.EntityDTO;
 public interface DenaDataStore {
     void storeObject(EntityDTO entityDTO);
 
-    EntityDTO findObject(long objectId);
+    EntityDTO findObject(Integer objectId);
 
 }
